Move AutoModerationRuleID to snowflake.go and go generate
